convert /holders to server-side initial data fetch
page.tsx is now an async server component that queries top_1000_holders
directly and passes rows to a new HoldersTable client component as
props. kills the post-hydration fetch and empty-table flash on first
paint. error path preserves the original inline message.
